Rough convertion of location menu component into rescript
